#168 moved some accomplishments helper functions to a new file to allow for decluttering and unit testing
