+++ src/accomplishments/accomplishments.js
@@ -1,5 +1,9 @@
 import AccomplishmentsStorage from "../backend-storage/accomplishments-api.js";
-
+import {
+    sortAccomplishments,
+    filterAccomplishments,
+    copyAccomplishments,
+} from "./accomplishments-helpers.js";
 /**
  * Get all accomplishment Objects from storage
  * @returns {accomplishemntsObj[]} array of accomplishments objects retrieved from storage
@@ -26,78 +30,6 @@ function displayMsgIfEmpty(allAccomplishmentsObj) {
 }
 
 /**
- * Sort an array of accomplishments objects by oldest or newest given an unsorted or sorted array
- * This allows for quick reversing betweeen newest and oldest rather than fully resorting each time
- * @param {accomplishemntsObj[]} accomplishmentsObjArr array of accomplishments objects
- * @param {Object} options contains byOldest (boolean) and alreadySortedByOldest(boolean)
- * @returns {accomplishemntsObj[]}
- */
-function sortAccomplishments(
-    accomplishmentsObjArr,
-    options = { byOldest: true, alreadySortedByOldest: true }
-) {
-    // create shallow copy of accomplishments to not edit the parameter that was passed in
-    let allAccomplishments = _copyAccomplishments(accomplishmentsObjArr);
-    // if it's already sorted how it's requested, return the shallow copy
-    if (options.byOldest && options.alreadySortedByOldest) {
-        return allAccomplishments;
-    }
-    // sort by oldest by default
-    if (!options.alreadySortedByOldest) {
-        allAccomplishments.sort(
-            (accomplishmentsObjOne, accomplishmentsObjTwo) => {
-                const dateOne = _getDate(accomplishmentsObjOne);
-                const dateTwo = _getDate(accomplishmentsObjTwo);
-                return dateOne.getTime() - dateTwo.getTime();
-            }
-        );
-    }
-    // sort by newest if requested
-    if (!options.byOldest) {
-        allAccomplishments = allAccomplishments.reverse();
-    }
-    return allAccomplishments;
-}
-
-/**
- * Filters an array of accomplishments objects by current year, current month, or no filtering
- * @param {accomplishemntsObj[]} accomplishmentsObjArr array of accomplishments objects
- * @param {Object} options contains byCurrentMonth (Boolean) and byCurrentYear (Boolean)
- * @throws {Error} if options has incorrect parameter settings
- * @returns {accomplishemntsObj[]}
- */
-function filterAccomplishments(
-    accomplishmentsObjArr,
-    options = { byCurrentMonth: false, byCurrentYear: false }
-) {
-    if (options.byCurrentMonth && options.byCurrentYear) {
-        throw Error("Cannot filter by year and month at the same time.");
-    }
-    // create shallow copy of accomplishments to not edit the parameter that was passed in
-    let filteredAccomplishments = _copyAccomplishments(accomplishmentsObjArr);
-    // filter by current month if requested
-    if (options.byCurrentMonth) {
-        const now = new Date();
-        const currentMonth = now.getMonth();
-        const currentYear = now.getFullYear();
-        filteredAccomplishments = filteredAccomplishments.filter(
-            (accomplishmentsObj) =>
-                _getDate(accomplishmentsObj).getMonth() === currentMonth &&
-                _getDate(accomplishmentsObj).getFullYear() === currentYear
-        );
-    }
-    // filter by current year if requested
-    if (options.byCurrentYear) {
-        const currentYear = new Date().getFullYear();
-        filteredAccomplishments = filteredAccomplishments.filter(
-            (accomplishmentsObj) =>
-                _getDate(accomplishmentsObj).getFullYear() === currentYear
-        );
-    }
-    return filteredAccomplishments;
-}
-
-/**
  * Populate the accomplishments table given an arry of accomplishments objects.
  * Puts every accomplishment on a new row.
  * @param {accomplishemntsObj[]} accomplishmentsObjArr array of accomplishments objects
@@ -105,7 +37,7 @@ function filterAccomplishments(
 function populateTable(accomplishmentsObjArr) {
     _clearTable();
     // create shallow copy of accomplishments to not edit the parameter that was passed in
-    let allAccomplishments = _copyAccomplishments(accomplishmentsObjArr);
+    let allAccomplishments = copyAccomplishments(accomplishmentsObjArr);
     // get table, add a row to table for each accomplishment in each accomplishments object
     const tableBody = document.getElementById("tableBody");
     for (const accomplishment of allAccomplishments) {
@@ -135,33 +67,11 @@ function populateTable(accomplishmentsObjArr) {
 }
 
 /**
- * Get the date of an accomplishments object as a Date object
- * @param {accomplishemntsObj} accomplishmentsObj an accomplishments object - has date and content fields
- * @returns {Date} Date object corresponding to the accomplishments object's date
- */
-function _getDate(accomplishmentsObj) {
-    return new Date(accomplishmentsObj.date);
-}
-
-/**
  * Clears the accomplishments table
  */
 function _clearTable() {
     const tableBody = document.getElementById("tableBody");
     tableBody.innerHTML = "";
-}
-
-/**
- * return a shallow copy of an array that still points to the accomplishments objects
- * @param {accomplishemntsObj[]} AccomplishmentsObjArr array of accomplishments objects
- * @returns {accomplishemntsObj[]} shallow copy of parameter
- */
-function _copyAccomplishments(AccomplishmentsObjArr) {
-    let copiedAccomplishmentsObj = [];
-    for (const accomplishemntsObj of AccomplishmentsObjArr) {
-        copiedAccomplishmentsObj.push(accomplishemntsObj);
-    }
-    return copiedAccomplishmentsObj;
 }
 
 /**
